push_swap norm done

+++ WIP/push_swap/utils/ft_bzero.c
@@ -6,13 +6,13 @@
 /*   By: echung <echung@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/21 20:21:56 by echung            #+#    #+#             */
-/*   Updated: 2021/06/27 16:14:29 by echung           ###   ########.fr       */
+/*   Updated: 2021/07/01 01:07:49 by echung           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-void	ft_bzero(void *s, size_t n)
+void	ft_bzero(void *s, int n)
 {
 	ft_memset(s, 0, n);
 }
