+++ WIP/push_swap/utils/ft_memset.c
@@ -6,13 +6,13 @@
 /*   By: echung <echung@student.42seoul.kr>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/21 20:23:02 by echung            #+#    #+#             */
-/*   Updated: 2021/06/27 16:14:10 by echung           ###   ########.fr       */
+/*   Updated: 2021/07/01 01:08:16 by echung           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "push_swap.h"
 
-void	*ft_memset(void *b, int c, size_t len)
+void	*ft_memset(void *b, int c, int len)
 {
 	unsigned char	*b_org;
 	unsigned char	*b_mod;
